build with -Wall fix associated warnings (at least on gcc 4.0.1)
git-svn-id: http://skia.googlecode.com/svn/trunk@129 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ tests/PackBitsTest.cpp
@@ -45,7 +45,7 @@ static void test_pack16(skiatest::Reporter* reporter) {
         size_t maxSize = SkPackBits::ComputeMaxSize16(size);
         REPORTER_ASSERT(reporter, maxSize >= dstSize);
 
-        int srcCount = SkPackBits::Unpack16(dst, dstSize, src2);
+        size_t srcCount = SkPackBits::Unpack16(dst, dstSize, src2);
         REPORTER_ASSERT(reporter, size == srcCount);
         bool match = memcmp(src, src2, size * sizeof(uint16_t)) == 0;
         REPORTER_ASSERT(reporter, match);
@@ -99,7 +99,7 @@ static void test_pack8(skiatest::Reporter* reporter) {
             size_t maxSize = SkPackBits::ComputeMaxSize8(size);
             REPORTER_ASSERT(reporter, maxSize >= dstSize);
 
-            int srcCount = SkPackBits::Unpack8(dst, dstSize, src2);
+            size_t srcCount = SkPackBits::Unpack8(dst, dstSize, src2);
             REPORTER_ASSERT(reporter, size == srcCount);
             bool match = memcmp(src, src2, size * sizeof(uint8_t)) == 0;
             REPORTER_ASSERT(reporter, match);
